# Improvement on cache invalidation