@cynecx 2020-01-11
Note that this will significantly increase memory usage of channels which is not really desirable (Since now a slot value’s size will be aligned to a multiple of 128 bytes, at least on x86-64).