@Licenser 2020-01-11
That's a good point, especially for small values the memory growth would be quite a bit, OTOH  especially for them the performance difference is significant too.

I'm not sure what the right trade off is, perhaps it'd be better suited as a own flavor.